+++ tests/test_06_math/test_617_clockwise_orientation.py
@@ -11,12 +11,20 @@ def v1():
     return [(0, 0), (1, 0), (1, 1), (0, 1)]
 
 
-def test_has_clockwise_orientation(v1):
-    assert has_clockwise_orientation(v1) is True
+@pytest.mark.parametrize('vertices', [
+    [(0, 0), (1, 0), (1, 1)],
+    [(0, 0), (1, 0), (1, 1), (0, 1)],
+])
+def test_has_clockwise_orientation(vertices):
+    assert has_clockwise_orientation(vertices) is False
 
 
-def test_has_counter_clockwise_orientation(v1):
-    assert has_clockwise_orientation(reversed(v1)) is False
+@pytest.mark.parametrize('vertices', [
+    [(1, 1), (1, 0), (0, 0)],
+    [(0, 1), (1, 1), (1, 0), (0, 0)],
+])
+def test_has_counter_clockwise_orientation(vertices):
+    assert has_clockwise_orientation(vertices) is True
 
 
 if __name__ == '__main__':